Split Dependency on ASP.Net into it's own project
