+++ SagaFlow.sln
@@ -20,6 +20,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SagaFlow.Interfaces", "Saga
 EndProject
 Project("{54A90642-561A-4BB1-A94E-469ADEE60C69}") = "SagaFlow.UI", "SagaFlow.UI\SagaFlow.UI.esproj", "{856E3710-0012-4650-BDB8-47124B95D1F4}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SagaFlow.AspNetCore", "SagaFlow.AspNetCore\SagaFlow.AspNetCore.csproj", "{0D9F9A01-5FDB-4FD8-96C2-534AD71E6BBA}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -46,6 +48,10 @@ Global
 		{856E3710-0012-4650-BDB8-47124B95D1F4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{856E3710-0012-4650-BDB8-47124B95D1F4}.Release|Any CPU.ActiveCfg = Debug|Any CPU
 		{856E3710-0012-4650-BDB8-47124B95D1F4}.Release|Any CPU.Build.0 = Debug|Any CPU
+		{0D9F9A01-5FDB-4FD8-96C2-534AD71E6BBA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0D9F9A01-5FDB-4FD8-96C2-534AD71E6BBA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0D9F9A01-5FDB-4FD8-96C2-534AD71E6BBA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0D9F9A01-5FDB-4FD8-96C2-534AD71E6BBA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
