+++ SagaFlow/Global.cs
@@ -1,0 +1,5 @@
+﻿
+using System.Runtime.CompilerServices;
+
+// Enables setting properties with internal setter in SagaFlow.Schema.Command and SagaFlow.Schema.ResourceProvider
+[assembly: InternalsVisibleTo("SagaFlow.AspNetCore")]
