+++ SagaFlow/ISagaFlowSchemaProvider.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using SagaFlow.Schema;
 
 namespace SagaFlow
